+++ .pre-commit-config.yaml
@@ -3,20 +3,20 @@ repos:
     rev: 19.10b0
     hooks:
       - id: black
-        language_version: python3.7
+        language_version: python3.8
         args: ["--safe"]
 
   - repo: https://github.com/PyCQA/isort
     rev: 5.4.2
     hooks:
       - id: isort
-        language_version: python3.7
+        language_version: python3.8
 
   - repo: https://github.com/PyCQA/flake8
     rev: 3.8.3
     hooks:
       - id: flake8
-        language_version: python3.7
+        language_version: python3.8
         args: [
             # E501 let black handle all line length decisions
             # W503 black conflicts with "line break before operator" rule
@@ -28,7 +28,7 @@ repos:
     rev: 5.1.1
     hooks:
       - id: pydocstyle
-        language_version: python3.7
+        language_version: python3.8
         args: [
             # Check for docstring presence only
             "--select=D1",
@@ -40,5 +40,5 @@ repos:
     rev: v0.770
     hooks:
       - id: mypy
-        language_version: python3.7
+        language_version: python3.8
         args: ["--no-strict-optional", "--ignore-missing-imports"]
