+++ tilebench/resources/__init__.py
@@ -1,0 +1,1 @@
+"""tilebench viz resources."""
