+++ tilebench/templates/index.html
@@ -382,7 +382,7 @@
                         html += `<tr><td class="align-l">LIST</td><td class="px3 align-r">${listCount}</td></tr>`
                         html += `<tr><td class="align-l">HEAD</td><td class="px3 align-r">${headCount}</td></tr>`
                         html += `<tr><td class="align-l">GET</td><td class="px3 align-r">${getCount}</td></tr>`
-                        html += `<tr><td class="align-l">GET (o)</td><td class="px3 align-r">${getSize}</td></tr>`
+                        html += `<tr><td class="align-l">GET (bytes)</td><td class="px3 align-r">${getSize}</td></tr>`
 
                         const timing = data.get("server-timing")
                         if (timing) {
